Pointing to the right file
So at first I wasn't pointing to the file correctly; then after I did, I realized that the file is simply too large for this site to process. So I took a subset of the data, just those stars within 200 lightyears or so, and cut out the others. This should hopefully be much easier to process and not run afoul of timeout limitations.