Making stars appear on the home page
I want the stars to be output on the home page in the final version, so I fixed the archive to point to home and made the endpoint jive with the slug rewrite. I also pointed to the API as the source instead of the URL, which I think is more proper? Not 100% sure.