removed stellar-classification-parser
Trying to use this library, kept running into issues using it because it was trying to scan a text file as some kind of glossary, and you just can't do that on the browser side. I looked at maybe extracting the parsing code without the I/O, but it just wasn't worth it. I'm going to go back and do this parsing on the PHP side of things.