Fixed import (again) & added mass/color codes
Import was only working if the database was empty, if full it only returned a "0", so I fixed that by actually thinking through it. I also added code to extract the luminosity class and color from the spectral type, and add a keyword to the database. That way, if I change from one color scheme to another (there are a few different interpretations of the data) I can do that easily in ReactJS and not have to update the SQL.